Fix problematic pointer arithmetic
Silences warning:
lightspark/src/backends/rendering.cpp: In member function ‘void lightspark::RenderThread::finalizeUpload()’:
lightspark/src/backends/rendering.cpp:126:37: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

+++ src/backends/rendering.h
@@ -47,7 +47,7 @@ private:
 	void commonGLDeinit();
 	GLuint pixelBuffers[2];
 	uint32_t currentPixelBuffer;
-	uint32_t currentPixelBufferOffset;
+	intptr_t currentPixelBufferOffset;
 	uint32_t pixelBufferWidth;
 	uint32_t pixelBufferHeight;
 	void resizePixelBuffers(uint32_t w, uint32_t h);
